+++ apparate/update_databricks_library.py
@@ -4,6 +4,8 @@ The core of the program, update_databricks_library handles all the logic and
  in Databricks.
 """
 import json
+import re
+from os.path import basename
 
 import requests
 from configparser import NoOptionError
@@ -37,7 +39,110 @@ class APIError(Exception):
         return '{}: {}'.format(self.code, self.message)
 
 
-def load_egg(filename, library_name, version, folder, token, host):
+class FileNameError(Exception):
+    """
+    exception to handle when filename is not of correct pattern
+    """
+    def __init__(self, filename):
+        Exception.__init__(
+            self,
+            'Filename \'{}\' was not correct pattern'.format(filename)
+        )
+        self.filename = filename
+
+
+class FileNameMatch(object):
+    """
+    Matches eggs or jars for both released and snapshot versions
+
+    Supported Patterns:
+      new_library-1.0.0-py3.6.egg
+      new_library-1.0.0-SNAPSHOT-py3.6.egg
+      new_library-1.0.0-SNAPSHOT-my-branch-py3.6.egg
+
+      new_library-1.0.0.egg
+      new_library-1.0.0-SNAPSHOT.egg
+      new_library-1.0.0-SNAPSHOT-my-branch.egg
+
+      new_library-1.0.0.jar
+      new_library-1.0.0-SNAPSHOT.jar
+      new_library-1.0.0-SNAPSHOT-my-branch.jar
+
+    library_name: string
+        base name of library (e.g. 'test_library')
+    version: string
+        version of library (e.g. '1.0.0')
+
+    """
+    file_pattern = (
+        r'([a-zA-Z0-9-\._]+)-((\d+)\.(\d+\.\d+)'
+        r'(?:-SNAPSHOT(?:[a-zA-Z_\-\.]+)?)?)(?:-py.+)?\.(egg|jar)'
+    )
+
+    def __init__(self, filename):
+        match = re.match(FileNameMatch.file_pattern, filename)
+        try:
+            self.filename = filename
+            self.library_name = match.group(1)
+            self.version = match.group(2)
+            self.major_version = match.group(3)
+            self.minor_version = match.group(4)
+            self.suffix = match.group(5)
+            if self.suffix == 'jar':
+                self.lib_type = 'java-jar'
+            elif self.suffix == 'egg':
+                self.lib_type = 'python-egg'
+        except (IndexError, AttributeError):
+            raise FileNameError(filename)
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            self_attrs = {k: v for k, v in vars(self).items()}
+            self_attrs.pop('filename')
+            other_attrs = {k: v for k, v in vars(other).items()}
+            other_attrs.pop('filename')
+            return self_attrs == other_attrs
+        else:
+            return False
+
+    def replace_version(self, other, logger):
+        """
+        True if self can safely replace other
+
+        based on version numbers only - snapshot and branch tags are ignored
+        """
+
+        if other.library_name != self.library_name:
+            logger.debug(
+                'not replacable: {} != {} ()'
+                .format(other.library_name, self.library_name, other.filename)
+            )
+            return False
+        elif int(other.major_version) != int(self.major_version):
+            logger.debug(
+                'not replacable: {} != {} ({})'
+                .format(
+                    int(self.major_version),
+                    int(other.major_version),
+                    other.filename,
+                )
+            )
+            return False
+        elif float(other.minor_version) >= float(self.minor_version):
+            logger.debug(
+                'not replacable: {} >= {} ({})'
+                .format(
+                    other.minor_version,
+                    self.minor_version,
+                    other.filename,
+                )
+            )
+            return False
+        else:
+            return True
+
+
+def load_library(filename, match, folder, token, host):
     """
     upload an egg to the Databricks filesystem.
 
@@ -45,10 +150,8 @@ def load_egg(filename, library_name, version, folder, token, host):
     ----------
     filename: string
         local location of file to upload
-    library_name: string
-        base name of library (e.g. 'test_library')
-    version: string
-        version of library (e.g. '1.0.0')
+    match: FilenameMatch object
+        match object with library_type, library_name, and version
     folder: string
         Databricks folder to upload to
         (e.g. '/Users/htorrence@shoprunner.com/')
@@ -61,31 +164,32 @@ def load_egg(filename, library_name, version, folder, token, host):
     ------------
     uploads egg to Databricks
     """
-    res = requests.post(
-        host + '/api/1.2/libraries/upload',
-        auth=('token', token),
-        data={
-            'libType': 'python-egg',
-            'name': '{0}-{1}'.format(library_name, version),
-            'folder': folder,
-        },
-        files={'uri': open(filename, 'rb')}
-    )
+    with open(filename, 'rb') as file_obj:
+        res = requests.post(
+            host + '/api/1.2/libraries/upload',
+            auth=('token', token),
+            data={
+                'libType': match.lib_type,
+                'name': '{0}-{1}'.format(match.library_name, match.version),
+                'folder': folder,
+            },
+            files={'uri': file_obj}
+        )
 
     if res.status_code != 200:
         raise APIError(res)
 
 
-def get_job_list(library_name, major_version, library_mapping, token, host):
+def get_job_list(logger, match, library_mapping, token, host):
     """
     get a list of jobs using the major version of the given library
 
     Parameters
     ----------
-    library_name: string
-        base name of library (e.g. 'test_library')
-    major_version: string
-        major version of library (e.g. '1')
+    logger: logging object
+        configured in cli_commands.py
+    match: FilenameMatch object
+        match object with suffix
     library_mapping: dict
         first element of get_library_mapping output
     token: string
@@ -107,41 +211,42 @@ def get_job_list(library_name, major_version, library_mapping, token, host):
         if len(res.json()['jobs']) == 0:
             return []
         for job in res.json()['jobs']:
+            logger.debug('job: {}'.format(job['settings']['name']))
             if 'libraries' in job['settings'].keys():
                 for library in job['settings']['libraries']:
-                    if 'egg' in library.keys():
+                    if match.suffix in library.keys():
                         try:  # if in prod_folder, mapping turns uri into name
-                            job_library_uri = library['egg'].split('/')[-1]
-                            job_library = library_mapping[job_library_uri]
+                            job_library_uri = basename(library[match.suffix])
+                            job_match = library_mapping[job_library_uri]
                         except KeyError:
+                            logger.debug(
+                                'not in library map: {}'
+                                .format(job_library_uri)
+                            )
                             pass
                         else:
-                            try:  # if we can parse a version, otherwise ignore
-                                job_library_name = (
-                                    '-'.join(job_library.split('-')[:-1])
-                                )
-                                job_major_version = (
-                                    job_library.split('-')[-1].split('.')[0]
-                                )
-                            except IndexError:
-                                pass
+                            if match.replace_version(job_match, logger):
+                                job_list.append({
+                                    'job_id': job['job_id'],
+                                    'job_name': job['settings']['name'],
+                                    'library_path': library[match.suffix],
+                                })
                             else:
-                                if (
-                                    (library_name == job_library_name)
-                                    and (job_major_version == major_version)
-                                ):
-                                    # append jobs we want to update later
-                                    job_list.append({
-                                        'job_id': job['job_id'],
-                                        'job_name': job['settings']['name'],
-                                        'library_path': library['egg'],
-                                    })
+                                logger.debug(
+                                    'not replacable: {}'
+                                    .format(job_match.filename)
+                                )
+                    else:
+                        logger.debug(
+                            'no matching suffix: looking for {}, found {}'
+                            .format(match.suffix, str(library.keys()))
+                        )
         return job_list
     else:
         raise APIError(res)
 
 
-def get_library_mapping(prod_folder, token, host):
+def get_library_mapping(logger, prod_folder, token, host):
     """
     returns a pair of library mappings, the first mapping library uri to a
      library name for all libraries in the production folder, and the second
@@ -150,6 +255,8 @@ def get_library_mapping(prod_folder, token, host):
 
     Parameters
     ----------
+    logger: logging object
+        configured in cli_commands.py
     prod_folder: string
         name of folder in Databricks UI containing production libraries
     token: string
@@ -183,24 +290,46 @@ def get_library_mapping(prod_folder, token, host):
             )
             if status_res.status_code == 200:
                 library_info = status_res.json()
-                # only do any of this for libraries in the ds folder
+                # only do any of this for libraries in the production folder
                 if library_info['folder'] != prod_folder:
-                    continue
-                library_map[library_info['files'][0]] = library_info['name']
-                try:
-                    name_split = library_info['name'].split('-')
-                    name_split = (
-                        ['-'.join(name_split[:-1])] + name_split[-1].split('.')
+                    logger.debug(
+                        'excluded folder: {} in {}, not prod folder ({})'
+                        .format(
+                            library_info['name'],
+                            library_info['folder'],
+                            prod_folder,
+                        )
                     )
-                    # map name to info
+                    continue
+                if library_info['libType'] == 'python-egg':
+                    full_name = library_info['name'] + '.egg'
+                elif library_info['libType'] == 'java-jar':
+                    full_name = library_info['name'] + '.jar'
+                else:
+                    logger.debug(
+                        'excluded library type: {} is of libType {}, '
+                        'not jar or egg'
+                        .format(
+                            library_info['name'],
+                            library_info['libType'],
+                        )
+                    )
+                    continue
+                try:
+                    name_match = FileNameMatch(full_name)
+                    # map uri to name match object
+                    library_map[library_info['files'][0]] = name_match
+                    # map name to name match object and id number
                     # we'll need the id number to clean up old libraries
                     id_nums[library_info['name']] = {
-                        'name': name_split[0],
-                        'major_version': name_split[1],
-                        'minor_version': '.'.join(name_split[2:]),
+                        'name_match': name_match,
                         'id_num': library_info['id'],
                     }
-                except IndexError:
+                except FileNameError:
+                    logger.debug(
+                        'FileNameError: {} file name is not parsable'
+                        .format(full_name)
+                    )
                     pass
             else:
                 raise APIError(status_res)
@@ -209,14 +338,25 @@ def get_library_mapping(prod_folder, token, host):
         raise APIError(res)
 
 
-def update_job_libraries(job_list, new_library_path, token, host):
+def update_job_libraries(
+    logger,
+    job_list,
+    match,
+    new_library_path,
+    token,
+    host,
+):
     """
     update libraries on jobs using same major version
 
     Parameters
     ----------
+    logger: logging object
+        configured in cli_commands.py
     job_list: list of strings
         output of get_job_list
+    match: FilenameMatch object
+        match object with suffix
     new_library_path: string
         path to library in dbfs (including uri)
     token: string
@@ -224,7 +364,6 @@ def update_job_libraries(job_list, new_library_path, token, host):
     host: string
         Databricks account url
         (e.g. https://fake-organization.cloud.databricks.com)
-
 
     Side Effects
     ------------
@@ -242,9 +381,12 @@ def update_job_libraries(job_list, new_library_path, token, host):
             job_specs.pop('settings')
             new_libraries = []
             for lib in settings['libraries']:
-                if 'egg' in lib.keys() and lib['egg'] == job['library_path']:
+                if (
+                    match.suffix in lib.keys()
+                    and lib[match.suffix] == job['library_path']
+                ):
                     # replace entry for old library path with new one
-                    new_libraries.append({'egg': new_library_path})
+                    new_libraries.append({match.suffix: new_library_path})
                 else:
                     new_libraries.append(lib)
             settings['libraries'] = new_libraries
@@ -261,9 +403,8 @@ def update_job_libraries(job_list, new_library_path, token, host):
 
 
 def delete_old_versions(
-    library_name,
-    major_version,
-    minor_version,
+    logger,
+    new_library_match,
     id_nums,
     token,
     prod_folder,
@@ -277,12 +418,10 @@ def delete_old_versions(
 
     Parameters
     ----------
-    library_name: string
-        base name of library (e.g. 'test_library')
-    major_version: string
-        major version of library (e.g. '1')
-    minor_version: string
-        minor version of library (e.g. '1.0')
+    logger: logging object
+        configured in cli_commands.py
+    match: FilenameMatch object
+        match object with library_name_, major_version, minor_version
     id_nums: dict
         second output of get_library_mapping
     token: string
@@ -301,12 +440,8 @@ def delete_old_versions(
 
     old_versions = []
     for name, lib in id_nums.items():
-        if (
-            lib['name'] == library_name
-            and int(lib['major_version']) == int(major_version)
-            and float(lib['minor_version']) < float(minor_version)
-        ):
-            old_versions.append(name)
+        if new_library_match.replace_version(lib['name_match'], logger):
+            old_versions.append(lib['name_match'].filename)
             res = requests.post(
                 host + '/api/1.2/libraries/delete',
                 auth=('token', token),
@@ -317,7 +452,7 @@ def delete_old_versions(
     return old_versions
 
 
-def update_databricks(path, token, folder, update_jobs, cleanup):
+def update_databricks(logger, path, token, folder, update_jobs, cleanup):
     """
     upload library, update jobs using the same major version,
     and delete libraries with the same major and lower minor versions
@@ -325,6 +460,8 @@ def update_databricks(path, token, folder, update_jobs, cleanup):
 
     Parameters
     ----------
+    logger: logging object
+        configured in cli_commands.py
     path: string
         path with name of egg as output from setuptools
         (e.g. dist/new_library-1.0.0-py3.6.egg)
@@ -360,66 +497,74 @@ def update_databricks(path, token, folder, update_jobs, cleanup):
         raise ValueError('no prod_folder provided: please run '
                          '`apparate configure` to get set up')
 
-    # strip off prepended folders
-    filename = path.split('/')[-1]
-    # strip off version and '-py3.6.egg'
-    library_name = '-'.join(filename.split('-')[:-2])
-    # get version
-    version = filename.split('-')[-2]
-    # get major version (e.g. 1) and minor version (e.g. 0.0)
-    major_version, minor_version = version.split('.', 1)
+    match = FileNameMatch(basename(path))
 
     try:
-        load_egg(path, library_name, version, folder, token, host)
-        print(
-            'new egg {}-{} loaded to Databricks'
-            .format(library_name, version)
+        load_library(path, match, folder, token, host)
+        logger.info(
+            'new library {}-{} loaded to Databricks'
+            .format(match.library_name, match.version)
         )
     except APIError as err:
         if err.code == 'http 500' and 'already exists' in err.message:
-            print(
-                'this version ({}) already exists:'.format(version) +
-                'if a change has been made please update your version number'
+            logger.info(
+                'This version ({}) already exists: '.format(match.version) +
+                'if a change has been made please update your version number. '
+                'Note this error can also occur if you are uploading a jar '
+                'and an egg already exists with the same name and version, '
+                'or vice versa. In this case you will need to choose a '
+                'different library name or a different folder for either the '
+                'egg or the jar.'
             )
             return
         else:
             raise err
 
     if update_jobs and folder == prod_folder:
-        library_map, id_nums = get_library_mapping(prod_folder, token, host)
+        library_map, id_nums = get_library_mapping(
+            logger,
+            prod_folder,
+            token,
+            host,
+        )
         library_uri = [
-            uri for uri, name in library_map.items()
-            if name == '{}-{}'.format(library_name, version)
+            uri for uri, tmp_match in library_map.items()
+            if (
+                match.library_name == tmp_match.library_name
+                and match.version == tmp_match.version
+            )
         ][0]
         library_path = 'dbfs:/FileStore/jars/' + library_uri
 
-        job_list = get_job_list(
-            library_name,
-            major_version,
-            library_map,
-            token,
-            host
-        )
-        print(
+        job_list = get_job_list(logger, match, library_map, token, host)
+        logger.info(
             'current major version of library used by jobs: {}'
             .format(', '.join([i['job_name'] for i in job_list]))
         )
 
         if len(job_list) != 0:
-            update_job_libraries(job_list, library_path, token, host)
-            print(
+            update_job_libraries(
+                logger,
+                job_list,
+                match,
+                library_path,
+                token,
+                host,
+            )
+            logger.info(
                 'updated jobs: {}'
                 .format(', '.join([i['job_name'] for i in job_list]))
             )
 
         if cleanup:
             old_versions = delete_old_versions(
-                library_name=library_name,
-                major_version=major_version,
-                minor_version=minor_version,
+                logger,
+                match,
                 id_nums=id_nums,
                 token=token,
                 prod_folder=prod_folder,
                 host=host,
             )
-            print('removed old versions: {}'.format(', '.join(old_versions)))
+            logger.info(
+                'removed old versions: {}'.format(', '.join(old_versions))
+            )